A little code cleanup
