+++ backend/exceptions/fs_exceptions.py
@@ -35,12 +35,3 @@ class RomAlreadyExistsException(Exception):
 
     def __repr__(self):
         return self.message
-
-
-class SchedulerException(Exception):
-    def __init__(self, message: str):
-        self.message = message
-        super().__init__(self.message)
-
-    def __repr__(self):
-        return self.message
